move applyDeclarativeNodeOptionParameters from NodeHelpers to DirectoryLoader
